Move UrlPaths and delete deprecated folder _common
