+++ src/app/views/developer/developer-data.service.ts
@@ -3,7 +3,7 @@ import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
 
 import { environment } from 'src/environments/environment';
-import { UrlPaths } from 'src/app/_common/constants/url-paths';
+import { UrlPaths } from 'src/app/url-paths';
 import { LibraryItem } from 'src/app/views/shared/library/models/library-item.model';
 
 @Injectable({
